Refactoring, comments
Added a few comments about headings nested inside elements (ie inside tables)
and malformed html. The short story is that the widget doesn't try to make those
right (there may not be such a thing as "right" anyway), the TOC will probably
be wrong and the navigation broken, but it shouldn't crash.